Resource based authorization

+++ Restaurants.Application/Dishes/Commands/DeleteDishes/DeleteDishesForRestautantCommandHandler.cs
@@ -1,7 +1,9 @@
 ﻿using MediatR;
 using Microsoft.Extensions.Logging;
+using Restaurants.Domain.Constants;
 using Restaurants.Domain.Entities;
 using Restaurants.Domain.Exceptions;
+using Restaurants.Domain.Interfaces;
 using Restaurants.Domain.Repositories;
 
 namespace Restaurants.Application.Dishes.Commands.DeleteDishes;
@@ -11,13 +13,16 @@ public class DeleteDishesForRestautantCommandHandler : IRequestHandler<DeleteDis
     private readonly ILogger<DeleteDishesForRestautantCommandHandler> _logger;
     private readonly IRestaurantsRepository _restaurantsRepository;
     private readonly IDishesRepository _dishesRepository;
+    private readonly IRestaurantAuthorizationService _restaurantAuthorizationService;
     public DeleteDishesForRestautantCommandHandler(ILogger<DeleteDishesForRestautantCommandHandler> logger,
                                                    IRestaurantsRepository restaurantsRepository,
-                                                   IDishesRepository dishesRepository)
+                                                   IDishesRepository dishesRepository,
+                                                   IRestaurantAuthorizationService restaurantAuthorizationService)
     {
         _logger = logger;
         _restaurantsRepository = restaurantsRepository;
         _dishesRepository = dishesRepository;
+        _restaurantAuthorizationService = restaurantAuthorizationService;
     }
     public async Task Handle(DeleteDishesForRestautantCommand request, CancellationToken cancellationToken)
     {
@@ -29,6 +34,13 @@ public class DeleteDishesForRestautantCommandHandler : IRequestHandler<DeleteDis
         {
             throw new ResourceNotFoundException(nameof(Restaurant), request.RestaurantId.ToString());
         }
+
+        bool IsDeleteOperationAuthorized = _restaurantAuthorizationService.Authorize(restaurant, ResourceOperation.Delete);
+        if (!IsDeleteOperationAuthorized)
+        {
+            throw new ForbidException();
+        }
+
 
         await _dishesRepository.Delete(restaurant.dishes);
     }
